requirements.txt
boto3==1.34.85
botocore==1.34.85
click==8.1.7
Flask==2.2.5
Flask-Cors==4.0.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
